requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
sqlalchemy>=2.0.25
alembic==1.13.1
